[SWM-90] Move : under '/src' folder
