--- app/components/ui/MainHeading.tsx
@@ -1,3 +1,0 @@
-export default function MainHeading({ content }: { content: string }) {
-  return <h1 className='text-6xl font-bold'>{content}</h1>;
-}
