--- app/page.tsx
@@ -1,7 +1,0 @@
-import MainHeading from "./components/ui/MainHeading";
-
-export default function Main() {
-  return <>
-  <MainHeading content={"유튜브 강의를 쉽고 편하게."}/>
-  </>;
-}
