+++ src/components/heading/MainDescription.tsx
@@ -1,0 +1,3 @@
+export default function MainDescription({ content }: { content: string }) {
+  return <h3 className='mt-5 ml-16 text-3xl font-semibold'>{content}</h3>;
+}
