+++ src/components/heading/MainHeading.tsx
@@ -1,0 +1,3 @@
+export default function MainHeading({ content }: { content: string }) {
+  return <h2 className='mt-24 ml-16 text-5xl font-bold mb-14'>{content}</h2>;
+}
